requirements.txt
einops
einops-exts
git+https://github.com/huggingface/transformers
torch
torchvision
pillow
more-itertools
datasets
braceexpand
webdataset
wandb
nltk
scipy
inflection
sentencepiece
open_clip_torch
omegaconf
iopath
opencv-python
decord
accelerate
bitsandbytes